refactor(constants): content-related, so put it there
