+++ website/common/script/constants.js
@@ -2,7 +2,6 @@ export const MAX_HEALTH = 50;
 export const MAX_LEVEL = 100;
 export const MAX_STAT_POINTS = MAX_LEVEL;
 export const ATTRIBUTES = ['str', 'int', 'per', 'con'];
-export const MAX_INCENTIVES = 100;
 
 export const TAVERN_ID = '00000000-0000-4000-A000-000000000000';
 export const LARGE_GROUP_COUNT_MESSAGE_CUTOFF = 5000;
