+++ website/common/script/content/constants.js
@@ -1,4 +1,5 @@
 /* eslint-disable key-spacing */
+export const MAX_INCENTIVES = 100;
 
 export const CLASSES = [
   'warrior',
