+++ website/common/script/content/loginIncentives.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { MAX_INCENTIVES } from '../constants';
+import { MAX_INCENTIVES } from './constants';
 
 module.exports = function getLoginIncentives (api) {
   let loginIncentives = {
