+++ website/common/script/index.js
@@ -21,14 +21,12 @@ import {
   MAX_HEALTH,
   MAX_LEVEL,
   MAX_STAT_POINTS,
-  MAX_INCENTIVES,
   TAVERN_ID,
   LARGE_GROUP_COUNT_MESSAGE_CUTOFF,
   SUPPORTED_SOCIAL_NETWORKS,
 } from './constants';
 
 api.constants = {
-  MAX_INCENTIVES,
   LARGE_GROUP_COUNT_MESSAGE_CUTOFF,
   SUPPORTED_SOCIAL_NETWORKS,
 };
